--- src/components/WritePost.jsx
@@ -1,42 +1,0 @@
-import React, { useState } from "react";
-import api from "../api";
-import { useNavigate } from "react-router-dom";
-
-function WritePost({ user }) {
-  const [title, setTitle] = useState("");
-  const [content, setContent] = useState("");
-  const navigate = useNavigate();
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      await api.post("/posts", { title, content });
-      navigate("/");
-    } catch (error) {
-      console.error("Error creating post:", error);
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <div>
-        <label>Title:</label>
-        <input
-          type="text"
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-        />
-      </div>
-      <div>
-        <label>Content:</label>
-        <textarea
-          value={content}
-          onChange={(e) => setContent(e.target.value)}
-        />
-      </div>
-      <button type="submit">Create Post</button>
-    </form>
-  );
-}
-
-export default WritePost;
