+++ src/pages/Home.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import api from "../api";
 import Post from "../components/Post";
+import "./Home.css";
 
 function Home() {
   const [posts, setPosts] = useState([]);
@@ -22,17 +23,17 @@ function Home() {
     fetchPosts();
   }, []);
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await api.post("/posts", { title, content });
-      setPosts([...posts, response.data]);
-      setTitle("");
-      setContent("");
-    } catch (error) {
-      console.error("Error creating post:", error);
-    }
-  };
+  // const handleSubmit = async (e) => {
+  //   e.preventDefault();
+  //   try {
+  //     const response = await api.post("/posts", { title, content });
+  //     setPosts([...posts, response.data]);
+  //     setTitle("");
+  //     setContent("");
+  //   } catch (error) {
+  //     console.error("Error creating post:", error);
+  //   }
+  // };
 
   if (loading) {
     return <div>Loading...</div>;
@@ -41,31 +42,13 @@ function Home() {
   return (
     <div>
       <h1>Board</h1>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <input
-            type="text"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-            placeholder="Title"
-            required
-          />
-        </div>
-        <div>
-          <textarea
-            value={content}
-            onChange={(e) => setContent(e.target.value)}
-            placeholder="Content"
-            required
-          />
-        </div>
-        <button type="submit">Create Post</button>
-      </form>
-      {posts.length === 0 ? (
-        <div>No posts available.</div>
-      ) : (
-        posts.map((post) => <Post key={post.id} post={post} />)
-      )}
+      <div className="card-container">
+        {posts.length === 0 ? (
+          <div>No posts available.</div>
+        ) : (
+          posts.map((post) => <Post key={post.id} post={post} />)
+        )}
+      </div>
     </div>
   );
 }
